geotiff tests: consolidate tile / strip decode cluster (#2429)
Fold seven top-level files into xrspatial/geotiff/tests/read/test_tiling.py:

- test_geotiff_planar_strip_truncation_1782.py
- test_mixed_bps.py
- test_mixed_sample_format.py
- test_parallel_decode_default_tile_1551.py
- test_parallel_strip_decode_2100.py
- test_planar_multiband.py
- test_unpack_bits_vectorised_1713.py

The existing byte-cap content (issues #1664, #2026-05-18) becomes
sections 1-2. Seven new sections sit in failure-mode order: mixed
BitsPerSample dispatch (#1505), mixed SampleFormat dispatch (#1868),
sub-byte unpack_bits vs reference (#1713), planar=2 strip table
truncation (#1782), planar / layout / band / dtype matrix
(audit A2 + A3), parallel tile decode gate (#1551), parallel strip
decode local + HTTP planar=1 / planar=2 (#2100).

Resolve helper duplication: the two minimal-IFD builders for mixed
BPS and SampleFormat share a single _finalize_minimal_ifd serialiser.
Class-style resolve_bits_per_sample / resolve_sample_format unit
suites stay grouped under TestResolveBitsPerSample /
TestResolveSampleFormat with descriptive ids.

CLUSTER_AUDIT_TILE_STRIP.md maps every old file::test to its new
file::test_id. Deleted in a final pre-merge commit.

Cluster 5 of long-tail epic #2424. Tests-only -- no source changes.
